begin theme build refactor
