+++ sandbox/grommet-app/src/themes/index.ts
@@ -1,0 +1,4 @@
+import { themes as oldThemes } from './theme';
+import { themes as newThemes } from './theme-new';
+
+export const themes = { ...oldThemes, ...newThemes };
