+++ sandbox/grommet-app/src/themes/theme-new.ts
@@ -1,0 +1,5 @@
+import { hpe as hpeV8 } from 'grommet-theme-hpe';
+
+export const themes = {
+  v8: hpeV8,
+};
